requirements.txt
# Core dependencies
pwntools>=4.8.0
requests>=2.28.0
flask>=2.2.0
flask-cors>=3.0.10

# AI integration
google-generativeai>=0.3.0

# Analysis tools
capstone>=4.0.2
unicorn>=2.0.0
ropgadget>=6.7

# Web interface
werkzeug>=2.2.0
jinja2>=3.1.0

# Utilities
python-magic>=0.4.27
pathlib2>=2.3.7